Move `teleport/lib/tdp` to `shared/libs/tdp`
