+++ web/packages/shared/libs/tdp/client.ts
@@ -101,7 +101,7 @@ interface TdpTransport {
 // sending client commands, and receiving and processing server messages. Its creator is responsible for
 // ensuring the websocket gets closed and all of its event listeners cleaned up when it is no longer in use.
 // For convenience, this can be done in one fell swoop by calling Client.shutdown().
-export default class Client extends EventEmitter {
+export class TdpClient extends EventEmitter {
   protected codec: Codec;
   protected socket: TdpTransport | undefined;
   private sdManager: SharedDirectoryManager;
