+++ web/packages/shared/libs/tdp/index.ts
@@ -1,6 +1,6 @@
 /**
  * Teleport
- * Copyright (C) 2023  Gravitational, Inc.
+ * Copyright (C) 2025 Gravitational, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -16,4 +16,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { PlayerClient } from './playerClient';
+export {
+  TdpClient,
+  TdpClientEvent,
+  useListener,
+  type BitmapFrame,
+} from './client';
+export * from './codec';
