+++ web/packages/teleport/src/DesktopSession/DesktopSession.story.tsx
@@ -24,10 +24,12 @@ import {
   makeProcessingAttempt,
   makeSuccessAttempt,
 } from 'shared/hooks/useAsync';
-
-import { TdpClient, TdpClientEvent } from 'teleport/lib/tdp';
-import { BitmapFrame } from 'teleport/lib/tdp/client';
-import { ClientScreenSpec } from 'teleport/lib/tdp/codec';
+import {
+  BitmapFrame,
+  ClientScreenSpec,
+  TdpClient,
+  TdpClientEvent,
+} from 'shared/libs/tdp';
 
 import { DesktopSession, DesktopSessionProps } from './DesktopSession';
